http://bugs.sakaiproject.org/jira/browse/OPC-50
Added messages to the bundle
Added the generation of SelectItems for the Forum organize
Added a list UIData parameter to the DiscussionForumTool so the iterator know the row
Added the generation of SelectItems for the Topics in a Forum for the Organize
Added the Organize button to the dfArea
added a new page to the area for organizing parts of message center
Added the faces-config routing to the new page
Changed the Forum and Topics to be sorted by the respective Sort Index
When saving topics and forums the code will reindex the sort index for any new topics or forums
When reading forums and topics the sort index is set if it needs it

To test:
Go to message center.  There is a new button for organizing discussion forums.  click the button.  The forums and topics should be ordered by creation date.  Each forum and topic should have an index starting at 1.  Change the index of the forums and topics.  As the you change the indices all other indices should automatically change to the correct value.  Clicking save setting.  going back to the forum main page, the forums should have the new order specified.

git-svn-id: https://source.sakaiproject.org/svn/msgcntr/trunk@17606 66ffb92e-73f9-0310-93c1-f5514f145a0a