SAK-11130 http://jira.sakaiproject.org/jira/browse/SAK-11130 With a localized Sakai, if pvt_deleted, pvt_received and pvt_sent are not set to "Deleted", "Received" and "Sent" (default values). The folders that are not set to default values do not work.
Without looking at the code I think those default folders names have a internal name which is the same as their default label. When the label is localized, the code is unable to match the localized label to the internal name.

Quick fix: do not localize pvt_deleted, pvt_received and pvt_sent. I've done it and it works.

Better but not-so-quick fix: match those default folders localized labels and internal names in the code. I won't complain if someone else does it. :-)

it support english and spanish now.

git-svn-id: https://source.sakaiproject.org/svn/msgcntr/trunk@37680 66ffb92e-73f9-0310-93c1-f5514f145a0a