PermissionsLevelManager was preventing the whole of Sakai from Starting up after the K1 migration.
I spent about 60 mins looking into it, but I couldnt work out exactly why.
So I have wrapped the init in an exception trap so the rest of Sakai can start.

I doubt that msg forums will work untill this is fixed, very sorry to have to do this, lack of time.



git-svn-id: https://source.sakaiproject.org/svn/msgcntr/trunk@49761 66ffb92e-73f9-0310-93c1-f5514f145a0a