SAK-25007 Don't attempt to create index twice.
On MFR_HIDDEN_GROUPS_T table there is the same index with different names, the upgrade script attempts to create 2 copies of the index instead (which fails). I'm fixing the upgrade script so it creates upgrades the database to match a newly created clean database and another issue should be created to remove the duplicate index.

git-svn-id: https://source.sakaiproject.org/svn/msgcntr/trunk@310297 66ffb92e-73f9-0310-93c1-f5514f145a0a